fix RandomRotate visualize bugs, the machenism is quite complex, here is a short note.
The visualization hook will fetch inputs and samples at higher level of call stack than `data_preprocessor`. During `data_preprocessor` 's forward, the input will be detached from the dict, and the rotate augmentation will not be related to the outer call stack due to this reason. On the other side, the visualization uses data_samples generated by model's `predict` function, which is generated by the rotated image.

As a result, the visualization hook will receive `data_batch` with original image and `data_samples` with sematic maps corresponding to the rotated image.

There seems to be no fix method which do not require a modification on OpenMIM framework for now, temporary HACK is to disable `data_preprocessor`'s augmentation during test or validate.

The root reason for this issue is the OpenMIM framework didn't provide an official augmentation API after `DataLoader` returns samples.